+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
 ### Deprecated
 ### Removed
+- Removed the MetricFactory class. Decorators still work in the same way but the Factory is no longer needed.
 ### Fixed
 
 ## [0.1.7] - 2018-08-14